requirements.txt
pandas>=2.2
SQLAlchemy>=2.0
pyodbc>=5.1
python-dotenv>=1.0
